Copy tests for convert break tags to a new module.
